+++ src/components/card/card.stories.tsx
@@ -22,7 +22,10 @@ Base.args = {
 // elevated card
 export const Elevated = (args: CardProps) => (
   <Card {...args} elevate>
-    <Image />
+    <Image
+      imageUrl='https://source.unsplash.com/8TQUF6UbpAk/1600x900'
+      imageAlt='woman in black and white crew neck long sleeve shirt wearing purple knit cap'
+    />
     <Text>John Doe</Text>
     <Button block pill size='large'>
       Follow
@@ -35,7 +38,10 @@ export const Bordered = (args: CardProps) => (
   <Flex>
     <Box>
       <Card {...args}>
-        <Image />
+        <Image
+          imageUrl='https://source.unsplash.com/8TQUF6UbpAk/1600x900'
+          imageAlt='woman in black and white crew neck long sleeve shirt wearing purple knit cap'
+        />
         <Text>John Doe</Text>
         <Button block pill size='large'>
           Follow
@@ -46,7 +52,10 @@ export const Bordered = (args: CardProps) => (
     <Spacer ml={40}>
       <Box>
         <Card {...args} bordered={false}>
-          <Image />
+          <Image
+            imageUrl='https://source.unsplash.com/8TQUF6UbpAk/1600x900'
+            imageAlt='woman in black and white crew neck long sleeve shirt wearing purple knit cap'
+          />
           <Text>John Doe</Text>
           <Button block pill size='large'>
             Follow